requirements.txt
boto3==1.26.136
celery==5.2.7
coverage==7.2.5
Django==3.2.19
django-dotenv==1.4.2
django-filter==23.2
django-phonenumber-field==7.1.0
django-split-settings==1.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
model-bakery==1.11.0
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
redis==4.5.5
whitenoise==6.4.0
## The following requirements were added by pip freeze:
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
botocore==1.29.165
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
filelock==3.12.2
identify==2.5.25
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.1
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.1
platformdirs==3.9.1
pre-commit==3.3.3
prompt-toolkit==3.0.39
psycopg2==2.9.6

python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.1
wcwidth==0.2.6
